mxnavi: "system" in a separate thread

+++ Navi/main.cpp
@@ -130,18 +130,47 @@ int main(int argc, char* argv[])
 
 			LogFile.Write(Cmd);
 
-			int ResCode = 0;
+			bool ThreadSystem_ExitOK = false;
+			std::thread ThreadSystem([&ThreadSystem_ExitOK, &LogFile, &Cmd]()
+				{
+					int ResCode = 0;
 #if defined(_WIN32)
-			std::cout << Cmd << '\n';
+					std::cout << Cmd << '\n';
 #else 
-			ResCode = system(Cmd.c_str());
+					ResCode = system(Cmd.c_str());
 #endif
-			LogFile.Write("ResCode: " + std::to_string(ResCode));
+					LogFile.Write("ResCode: " + std::to_string(ResCode));
+					ThreadSystem_ExitOK = true;
+				});
 
-			for (auto& i : PictList)
+
+			//[#] 500ms * 20 = 10000ms => 10s waits for sending email
+			for (int i = 0; i < 20; ++i)
 			{
-				std::filesystem::remove(i);
+				std::this_thread::sleep_for(std::chrono::milliseconds(500));
+				if (ThreadSystem_ExitOK)
+				{
+					ThreadSystem.join();
+					break;
+				}
 			}
+
+			if (!ThreadSystem_ExitOK)
+			{
+				std::cerr << "[ERR] \"system\" failed\n";
+				// The program is being closed without joining existing thread.
+				exit(static_cast<int>(utils::tExitCode::EX_IOERR));
+				// If "return" is called while the program is being terminated without joining existing thread.
+				// The following message appears:
+				// "terminate called without an active exception
+				// Aborted(core dumped)"
+				//return static_cast<int>(utils::tExitCode::EX_IOERR);
+			}
+
+			// Do not try to remove them in case of trouble with the "system" thread.
+			// They will be removed at next start of this application.
+			for (auto& i : PictList)
+				std::filesystem::remove(i);
 		}
 	}
 	catch (const std::exception& e)
